+++ lib/prefetch.js
@@ -1,7 +1,7 @@
 import {ModelMap, ResourceKeys} from './config';
 import {noOp, once} from './utils';
 
-import {getCacheKey} from './index';
+import {getCacheKey} from './resourcerer';
 import request from './request';
 
 // prevents api request bursts based on quick swipes
