+++ lib/resourcerer.js
@@ -1049,6 +1049,3 @@ function fetchResources(resources, props, {
     })
   );
 }
-
-export {default as prefetch} from './prefetch';
-export {Model, Collection};
